Delete Record Modification done

+++ admin/includes/queries/delete_multiple_visitors.php
@@ -7,11 +7,19 @@ $data = json_decode(file_get_contents("php://input"), true);
 if ($data['type'] == 'current') {
 
     $delete_data = $data['delete_data'];
+    // print_r($delete_data);
     foreach ($delete_data as $key => $val) {
 
-        // print_r($delete_data);
-        $sql = "DELETE from visitors where BlockNumber='" . $val['block'] . "' AND FlatNumber=" . $val['flatno'] . " AND VisitorName='" . $val['vname'] . "';";
-        mysqli_query($con, $sql);
+        $enddate = date('Y-m-d', strtotime($val['startdate']. ' + ' . $val['duration'] .'day'));
+        $todaysDate = date("Y-m-d");
+    
+        if (strtotime($enddate) > strtotime($todaysDate)){
+            $sql = "DELETE from visitors where BlockNumber='" . $val['block'] . "' AND FlatNumber=" . $val['flatno'] . " AND VisitorName='" . $val['vname'] . "';";
+            // echo '<script>console.log('.$sql.')</script>';
+            mysqli_query($con, $sql);
+            
+        }
+
     }
 
 }
